반응형 그리그 구현 및 folder컴포넌트 생성
