--- src/App.css
@@ -1,71 +1,0 @@
-@import url(./utils/color.css);
-@import url(./utils/font.css);
-
-* {
-  box-sizing: border-box;
-}
-
-html {
-  font-size: 16px;
-}
-
-p {
-  margin: 0;
-}
-
-.body {
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-family: Pretendard;
-}
-
-.header {
-  background-color: var(--Background);
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding: 20px 0 60px 0;
-}
-
-.header img {
-  width: 60px;
-  height: 60px;
-}
-
-.ownerName {
-  font-family: Pretendard;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 24px; /* 150% */
-  margin-top: 12px;
-}
-
-.folderName {
-  color: #000;
-  text-align: center;
-  font-feature-settings: "clig" off, "liga" off;
-  font-family: Pretendard;
-  font-size: 40px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  margin-top: 20px;
-}
-
-.mainContent {
-  margin: 0 auto;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-
-.contentContainer {
-  gap: 20px;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-}
