--- src/components/part2-weekly-mission.code-workspace
@@ -1,8 +1,0 @@
-{
-  "folders": [
-	{
-		"path": "../.."
-	}
-],
-  "settings": {}
-}
